Fixed unit tests and code smells
