+++ src/main/controllers/features/unincorporated/businessAddressAutoLookupController.ts
@@ -1,10 +1,10 @@
 import { Session } from "@companieshouse/node-session-handler";
 import { NextFunction, Request, Response } from "express";
-import { validationResult } from "express-validator";
+import { ValidationError, validationResult } from "express-validator";
 import { BUSINESS_NAME } from "../../../common/__utils/constants";
 import * as config from "../../../config";
 import { AddressLookUpService } from "../../../services/address/addressLookUp";
-import { getAddressFromPostcode, getUKAddressesFromPostcode } from "../../../services/postcode-lookup-service";
+import { getAddressFromPostcode } from "../../../services/postcode-lookup-service";
 import { BASE_URL, UNINCORPORATED_BUSINESS_ADDRESS_CONFIRM, UNINCORPORATED_BUSINESS_ADDRESS_LIST, UNINCORPORATED_BUSINESS_ADDRESS_MANUAL, UNINCORPORATED_WHICH_SECTOR } from "../../../types/pageURL";
 import { addLangToUrl, getLocaleInfo, getLocalesService, selectLang } from "../../../utils/localise";
 import { FormattedValidationErrors, formatValidationError } from "../../../validation/validation";
@@ -57,8 +57,14 @@ export const post = async (req: Request, res: Response, next: NextFunction) => {
                     const nextPageUrl = addLangToUrl(BASE_URL + UNINCORPORATED_BUSINESS_ADDRESS_LIST, lang);
                     res.redirect(nextPageUrl);
                 }
-            }).catch((error) => {
-                const pageProperties = getPageProperties(formatValidationError(error, lang));
+            }).catch(() => {
+                const validationError : ValidationError[] = [{
+                    value: postcode,
+                    msg: "correspondenceLookUpAddressInvalidAddressPostcode",
+                    param: "postcode",
+                    location: "body"
+                }];
+                const pageProperties = getPageProperties(formatValidationError(validationError, lang));
                 res.status(400).render(config.UNINCORPORATED_BUSINESS_ADDRESS_LOOKUP, {
                     previousPage: addLangToUrl(BASE_URL + UNINCORPORATED_WHICH_SECTOR, lang),
                     title: "What is your business address?",
