+++ src/main/services/postcode-lookup-service.ts
@@ -1,9 +1,8 @@
 import { UKAddress } from "@companieshouse/api-sdk-node/dist/services/postcode-lookup";
 import { createAndLogError, logger } from "../utils/logger";
-import { Resource, createApiClient } from "@companieshouse/api-sdk-node";
+import { Resource } from "@companieshouse/api-sdk-node";
 import ApiClient from "@companieshouse/api-sdk-node/dist/client";
 import { createPublicApiKeyClient } from "./api-services";
-import { ValidationError } from "express-validator";
 import { POSTCODE_ADDRESSES_LOOKUP_URL } from "../utils/properties";
 
 export const getUKAddressesFromPostcode = async (postcodeAddressesLookupURL: string, postcode: string): Promise<UKAddress[]> => {
@@ -35,14 +34,8 @@ export const getIsValidUKPostcode = async (postcodeValidationUrl: string, postco
 export async function getAddressFromPostcode (postcode: string) {
     postcode = postcode.replace(/\s/g, "");
     const ukAddresses = await getUKAddressesFromPostcode(POSTCODE_ADDRESSES_LOOKUP_URL, postcode);
-    const validationError : ValidationError[] = [{
-        value: postcode,
-        msg: "correspondenceLookUpAddressInvalidAddressPostcode",
-        param: "postcode",
-        location: "body"
-    }];
     if (!ukAddresses.length) {
-        throw validationError;
+        throw Error("Postcode not found");
     }
     return ukAddresses;
 }
